requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
rich==13.7.1
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
